requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pytesseract
pillow
pdfplumber
python-multipart
reportlab